rename headsearch to footer
