+++ src/functions/filter-region-footer.php
